+++ Develappers.BillomatNet/InvoiceService.cs
@@ -609,6 +609,26 @@ namespace Develappers.BillomatNet
         }
 
         /// <summary>
+        /// Deletes an invoice payment.
+        /// </summary>
+        /// <param name="id">The ID of the invoice payment.</param>
+        /// <param name="token">The cancellation token.</param>
+        /// <returns>
+        /// A task that represents the asynchronous operation.
+        /// </returns>
+        /// <exception cref="ArgumentException">Thrown when the parameter check fails.</exception>
+        /// <exception cref="NotAuthorizedException">Thrown when not authorized to access this resource.</exception>
+        /// <exception cref="NotFoundException">Thrown when the resource url could not be found.</exception>
+        public Task DeletePaymentAsync(int id, CancellationToken token = default)
+        {
+            if (id <= 0)
+            {
+                throw new ArgumentException("invalid invoice payment id", nameof(id));
+            }
+            return DeleteAsync($"/api/invoice-payments/{id}", token);
+        }
+
+        /// <summary>
         /// Retrieves the invoice tag cloud.
         /// </summary>
         /// <param name="token">The cancellation token.</param>